trying multiple exportes components & some logic
For some reason notheing worked initially as it is here now, but then it did.
The 'Item' function is used by the 'PackingList' one with only the latter
needing 'export' and not the former. It uses it the logic withing Greetings.jsx
and then send results to main.jsx for rendering?