everything with '2',like App2, is sandbox project
Rather than joining snadbox or whatever it was, in odin project, to fork
the exercise, just copy n pasted it here with '2'.
Logic had issues with:
tried to make sure the counter goes up when a color is changed AND NOT
when a button is clicked. So if someone click greens a few times ina row
, which wouldn't change color, counter doesn't go up. For that set up an
colorArr, and was comparing new color to the last one in colorArr, but
last one in colorArr is actually second from last, as present color used
for logic as argument in the function here, was actually the one pushed
into colorArr as last, was doing 'length-1' for ages, didn't work, cause
then I acutally compared present to present, so "!==" logic doesn't work