trying to work out if can edit prev commit msgs
if not, forgot to add that instead of 'lenght-1' changed to 'lenght-2'
that worked, as last item pushed into colorArr is actially present item/
color which was used for present logic.

JUST CHECJED, you CAN change most recent msg, just like this is it,
"git commit --amend", but if you want to edit older than last ones,
it's some unpleasant force stuff, so not recommended.